% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{growthcurve}
\alias{growthcurve}
\title{Draw growth cruve from a fitted and sapmled nim object}
\usage{
growthcurve(f, ribbon_1_probs = c(0.05, 0.95), ribbon_2_probs = c(0.25,
  0.75), use_plotly = TRUE)
}
\arguments{
\item{f}{fitted and sapmled nim object}

\item{ribbon_1_probs}{probabilites interval of the 1st ribbon}

\item{ribbon_2_probs}{probabilites interval of the 2nd ribbon}

\item{use_plotly}{argument specifying whether to use/or not the 'plotly' package for interactive graph}
}
\description{
Draw growth cruve from a fitted and sapmled nim object
}
\examples{
data('precip_max')
extremes(precip_max) = 2:ncol(precip_max)
n <- nim( ~1, data = precip_max)
smp <- sample(n, length = 50)
f <- fit(n, smp, mc.cores = 4)
growthcurve(f)
growthcurve(f, c(.10, .90), c(.40,.60))
}

