% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gumbelplot}
\alias{gumbelplot}
\title{Draw gumble plot from a nim object}
\usage{
gumbelplot(nim, use_plotly = TRUE)
}
\arguments{
\item{nim}{nim object}

\item{use_plotly}{argument specifying whether to use/or not the 'plotly' package for interactive graph}
}
\description{
Draw gumble plot from a nim object
}

