% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nim.R
\docType{package}
\name{nim-package}
\alias{nim-package}
\title{nim: A package for fitting and assessment of non-stationary index floods models}
\description{
The nim package provides functions for fitting, uncertainty and goodness-of-fit assessment of Non-stationary Index-flood Models (nims) with smooth or parametric trends in the parameters of the GEV model together with some classes and methods for easy manipulation and assessment of results.
}
\author{
Martin Hanel \email{hanel@fzp.czu.cz}
}
\references{
Hanel, M., Buishand, T. A., Ferro, C. A. T. (2009)  A non-stationary index-flood model for precipitation extremes in transient Regional Climate Model simulations, Journal of Geophysical Research., 114(D15107). \href{http://onlinelibrary.wiley.com/doi/10.1029/2009JD011712/abstract}{link}
}

