% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nim.R
\name{nim}
\alias{nim}
\title{Fitting the non-stationary index-flood models}
\usage{
nim(formula = ~1, data, method = "Nelder-Mead", tol = 0.1, ...)
}
\arguments{
\item{formula}{one or two sided formula describing the model, see Details}

\item{data}{data.frame including the covariate and the extremes to be fitted}

\item{method}{method used for the parameter optimization see ?optim}

\item{tol}{minimum improvement of the log-likelihood between the iteration steps, if the difference between two succesive iteration steps is lower, the iteration ends}

\item{...}{other parameters passed to \code{optim(control = list(...))}}
}
\value{
Object of class \link{nim} - i.e. simply
}
\description{
Fitting the non-stationary index-flood models
}
\section{Details}{

The function is a simple wrapper for fitting non-stationary index-flood models with parametric (currently only linear) and smoothing trends in the parameters of the GEV model. Stationary index-flood models can be fitted also. For the data (\code{data.frame}) that are used within the function, the attribute \code{extremes} (see \link{extremes}) has to be set. The trend can be either smooth (\link{s}) or parametric (\link{p})n.
}

\section{See also}{

\link{sample}, \link{fit}, ...
}
\examples{
data('precip_max')
head(precip_max)
# indicate which columns contain the extremes to be fitted
extremes(precip_max) = 2:ncol(precip_max)

# stationary model
nim( ~1, data = precip_max)

# smooth trend in xi, with bandwidth h = 0.2
nim(xi ~ s(YR, h = 0.2), data = precip_max)

# smooth trend in all parameters, default bandwidth
nim(xi + g + k ~ s(YR), data = precip_max)

# linear trend in xi and gamma
nim(xi + g ~ p(YR), data = precip_max)
}

