% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{nims}
\alias{nims}
\title{Create nims object.}
\usage{
nims(...)
}
\arguments{
\item{...}{}
}
\value{
`nims` object
}
\description{
The function concatenates nim objects (typically the result of bootstrapping) into a nims object. Various helper functions are provided for comparison of various characteristics of individual nims.
}
\examples{

}

