% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sample.nim}
\alias{sample.nim}
\title{Create bootstrap sample based on fitted nim}
\usage{
\method{sample}{nim}(nim, length = 1, type = "parametric_average_cor",
  impute_NA = TRUE)
}
\arguments{
\item{nim}{fitted model}

\item{length}{number of bootstrap samples}

\item{type}{one of \code{parametric_average_cor} (the default), \code{parametric} and \code{nonparametric}}
}
\value{
list of bootstrap samples
}
\description{
Create bootstrap sample based on fitted nim
}

